[MOV] survey: prepare files for refactoring of survey_session
Move 'survey_session_chart.js', 'survey_session_colors.js',
'survey_session_leaderboard.js', 'survey_session_manage.js', and
'survey_session_text_answers.js' from '/src/js' to '/src/interactions'.
This commit is added to make it easier to follow the history.

task-4367641
