+++ addons/survey/static/src/interactions/survey_session_colors.js
