+++ addons/survey/static/src/interactions/survey_session_leaderboard.js
@@ -1,6 +1,6 @@
 import { rpc } from "@web/core/network/rpc";
 import publicWidget from "@web/legacy/js/public/public_widget";
-import SESSION_CHART_COLORS from "@survey/js/survey_session_colors";
+import SESSION_CHART_COLORS from "@survey/interactions/survey_session_colors";
 
 publicWidget.registry.SurveySessionLeaderboard = publicWidget.Widget.extend({
     init: function (parent, options) {
